+++ R/conventions.R
@@ -235,14 +235,15 @@
     # popPK models for amifampridine in patients with Lambert-Eaton
     # myasthenia (Thakkar 2017 doi:10.1002/psp4.12218).
     "acdap",
-    # H4 (active thiol) metabolite of clopidogrel: the pharmacologically
-    # active species responsible for P2Y12 receptor inhibition, formed
-    # via sequential CYP-mediated oxidation of clopidogrel (CYP2C19 is
-    # the dominant isoform for the second oxidation step). The "H4"
-    # designator refers to the H4 stereoisomer specifically, which is
-    # the antiplatelet-active diastereomer. Used in parent-plus-
-    # metabolite popPK models (Danielak 2017 doi:10.1007/s00228-017-2334-z).
-    "h4"
+    # Ritonavir (RTV), an HIV protease inhibitor that is co-administered
+    # at low dose as a pharmacokinetic booster of other HIV protease
+    # inhibitors (atazanavir, lopinavir, darunavir, ...) via potent
+    # CYP3A4 inhibition. Not a metabolite of the boosted PI but a
+    # co-administered perpetrator whose plasma concentration drives a
+    # direct inhibitory effect on the boosted PI's apparent clearance;
+    # the suffix marks the non-parent species in 2-drug simultaneous
+    # PK models (Schipani 2013 doi:10.1097/QAI.0b013e3182737231).
+    "rtv"
   ),
   # Suffixes allowed for multi-component CL parameters. `_ss` denotes
   # the steady-state arm; `_time` denotes the time-varying decay arm.